Changed int types to smaller

+++ main.h
@@ -1,7 +1,7 @@
 #ifndef __MAIN_H_
 #define __MAIN_H_
 
-void init(void* space, int dataSize);
+void init(void* space, short dataSize);
 void dispatch(void);
 
 #endif
